+++ FM24C.h
@@ -1,13 +1,9 @@
 /*!\file FM24C.h
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2023, SMFSW
-** \brief FRAM / EEPROM Driver (bank switching at I2C address level protocol)
-** \note The driver handles writing specificities for EEPROM type components
-** \note The driver is fully compatible with FRAM type components
-** \note When EEPROM / FRAM compatibility is not needed, FM24C_WRITE_SIZE can be set to FM24C_BANK_SIZE for more efficiency
+** \brief FM24C FRAM Driver (bank switching at I2C address level protocol)
 ** \details FM24C16B: 16-Kbit (2K * 8) Serial I2C F-RAM
 **			FM24C04B: 4-Kbit (512 * 8) Serial I2C F-RAM
-**			...
 ** \note	Compatibility (tested):
 **				- FM24C16B
 **				- FM24C04B
@@ -39,24 +35,17 @@
 // *****************************************************************************
 #ifndef FM24C_BASE_ADDR
 //! \note Define FM24C_BASE_ADDR to change default device base address
-#define FM24C_BASE_ADDR		0x50				//!< FM24C Base address
+#define FM24C_BASE_ADDR			0x50					//!< FM24C Base address
 #endif
 
-#define	FM24C_BANK_SIZE		0x100				//!< FM24C bank size (in bytes)
+#define	FM24C_BANK_SIZE			0x100					//!< FM24C bank size (in bytes)
 
-#define	FM24C04B_SIZE		0x100				//!< FM24C04B size (in bytes)
-#define	FM24C16B_SIZE		0x800				//!< FM24C size (in bytes)
-
-#define EEP_WRITE_SIZE		0x10				//!< Write buffer size (used for write procedures)
+#define	FM24C04B_SIZE			0x100					//!< FM24C04B size (in bytes)
+#define	FM24C16B_SIZE			0x800					//!< FM24C size (in bytes)
 
 #ifndef FM24C_SIZE
 //! \note FM24C_SIZE defaults to FM24C16B size, but can be defined to FM24C04B_SIZE (for example)
-#define FM24C_SIZE			FM24C16B_SIZE		//!< FM24C size
-#endif
-
-#ifndef FM24C_WRITE_SIZE
-//! \note FM24C_WRITE_SIZE defaults to EEP_WRITE_SIZE, but can be defined to FM24C_BANK_SIZE (in case of FRAM, no use of EEPROM write buffer)
-#define FM24C_WRITE_SIZE	EEP_WRITE_SIZE		//!< EEPROM type page size used for compatibility with both EEPROM/FRAM ships
+#define FM24C_SIZE				FM24C16B_SIZE			//!< FM24C size
 #endif
 
 
@@ -69,7 +58,6 @@
 typedef struct FM24C_t {
 	struct {
 	I2C_slave_t *		slave_inst;		//!< Slave structure
-	size_t				write_size;		//!< Useful for EEPROM writes (FRAM not restricted, can be changed to FM24C_BANK_SIZE)
 	PeripheralGPIO_t	WP_GPIO;		//!< Write Protect GPIO struct
 	} cfg;
 } FM24C_t;
@@ -114,7 +102,7 @@ FctERR FM24C_Init_Single(void);
 ** \param[in] nb - Number of bytes to read
 ** \return FctERR - error code
 **/
-FctERR NONNULL__ FM24C_Write(FM24C_t * const pCpnt, const uint8_t * data, const uint16_t addr, const uint16_t nb);
+FctERR NONNULL__ FM24C_Write(FM24C_t * const pCpnt, const uint8_t * const data, const uint16_t addr, const uint16_t nb);
 
 
 /*!\brief I2C Read function for FM24C
@@ -125,7 +113,7 @@ FctERR NONNULL__ FM24C_Write(FM24C_t * const pCpnt, const uint8_t * data, const 
 ** \param[in] nb - Number of bytes to read
 ** \return FctERR - error code
 **/
-FctERR NONNULL__ FM24C_Read(FM24C_t * const pCpnt, uint8_t * data, const uint16_t addr, const uint16_t nb);
+FctERR NONNULL__ FM24C_Read(FM24C_t * const pCpnt, uint8_t * const data, const uint16_t addr, const uint16_t nb);
 
 
 /****************************************************************/