+++ FM24C_ex.c
@@ -1,17 +1,13 @@
 /*!\file FM24C_ex.c
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2023, SMFSW
-** \brief FRAM / EEPROM Driver (bank switching at I2C address level protocol) extensions
-** \note The driver handles writing specificities for EEPROM type components
-** \note The driver is fully compatible with FRAM type components
-** \note When EEPROM / FRAM compatibility is not needed, FM24C_WRITE_SIZE can be set to FM24C_BANK_SIZE for more efficiency
+** \brief FM24C FRAM Driver (bank switching at I2C address level protocol) extensions
 ** \details FM24C16B: 16-Kbit (2K * 8) Serial I2C F-RAM
 **			FM24C04B: 4-Kbit (512 * 8) Serial I2C F-RAM
 ** \note	Compatibility (tested):
 **				- FM24C16B
 **				- FM24C04B
 **				- BR24T04FVM
-**				...
 **/
 /****************************************************************/
 #include "FM24C.h"
@@ -21,14 +17,11 @@
 /****************************************************************/
 
 
-#define VAL_CLR	0xFF	//!< Mass erase value for FM24C
-
-
 FctERR NONNULL__ FM24C_Mass_Erase(FM24C_t * const pCpnt)
 {
 	FctERR	err = ERROR_OK;
 	uint8_t	array[FM24C_BANK_SIZE];
-	memset(array, VAL_CLR, sizeof(array));
+	memset(array, FM24C_CLR_VAL, sizeof(array));
 
 	for (int i = 0 ; i < (FM24C_SIZE / sizeof(array)) ; i++)
 	{