+++ MB85RC256V_proc.h
@@ -1,13 +1,11 @@
 /*!\file MB85RC256V_proc.h
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2023, SMFSW
-** \brief MB85RC256V Driver procedures
+** \brief MB85RC256V FRAM Driver procedures
 ** \details MB85RC256V: 256-Kbit (32K * 8) I2C Memory FRAM
 ** \note	Compatibility (tested):
 **				- MB85RC256V
 **				- FM24CL64B
-** \note	Compatibility:
-**				- other components using same i2c protocol may be compatible
 **/
 /****************************************************************/
 #ifndef __MB85RC256V_PROC_H__